Removed menutext.h, renamed menutext2.h, unified refreshmenu ROM_M_3A: 7813 → 7788 - diff:  25
Total size reduction: 25 bytes
