--- SAUCE/defines/menutext.h
@@ -1,33 +1,0 @@
-
-#include "charmaps/bg_charmap.h"
-
-// const unsigned char menutext1[]="GEOMETRY DASH";
-// const unsigned char menutext2[]="PRESS START TO LOAD LEVEL ";
-
-
-// const unsigned char easytop[]="ab";
-// const unsigned char easybottom[]="cd";
-// const unsigned char normaltop[]="ef";
-// const unsigned char normalbottom[]="gh";
-// const unsigned char mediumtop[]="ij";
-// const unsigned char mediumbottom[]="kl";
-// const unsigned char hardtop[]="mn";
-// const unsigned char hardbottom[]="op";
-// const unsigned char insanetop[]="qr";
-// const unsigned char insanebottom[]="st";
-
-
-
-
-const unsigned char completetext[]="*";
-const unsigned char blanktext[]="*";
-const unsigned char blubtext[]=" ";
-
-
-
-const unsigned char menubgcolors[]={
-    0x00
-};
-
-#include "charmaps/color1_charmap.h"
-
